Added confirm password in Register Added show password option in Register Formatted code in Login

+++ frontend-react/src/pages/Register.js
@@ -13,11 +13,15 @@ class Register extends Component {
             showpassword: 'password',
 
             // Input Validation
-            formErrors: { username: '', name: '', email: '', password: '', address: '', mobile: '' },
+            formErrors: {
+                username: '', name: '', email: '', password: '',
+                confirmpassword: '', address: '', mobile: ''
+            },
             usernameValid: false,
             nameValid: false,
             emailValid: false,
             passwordValid: false,
+            confirmPasswordValid: false,
             addressValid: false,
             mobileValid: false,
             formValid: false
@@ -26,26 +30,26 @@ class Register extends Component {
     }
 
     registerClicked = (event) => {
-        event.preventDefault()
         let username = this.state.username;
         let name = this.state.name
         let email = this.state.email;
         let password = this.state.password;
+        let confirmpassword = this.state.confirmpassword;
         let address = this.state.address;
         let mobile = this.state.mobile;
-        let authHeader = window.btoa(username + ':' + email + ':' + password + ':' + address + ':' + mobile);
+        let authHeader = window.btoa(username + ':' + name + ':' + email + ':' + password + ':' + confirmpassword + ':' + address + ':' + mobile);
         let user = { 'username': username, 'authHeader': authHeader };
-        AuthService.register(username, email, password, name, address, mobile)
-        .then(response => {
-            console.log(response)
-            AuthService.signin(username, password)
-            localStorage.setItem('user', JSON.stringify(user));
-            this.props.history.push('/products')
-        })
-        .catch(response => {
-            console.log(response)
-            this.props.history.push('/register')
-        })
+        AuthService.register(username, name, email, password, confirmpassword, address, mobile)
+            .then(response => {
+                console.log(response)
+                AuthService.signin(username, password)
+                localStorage.setItem('user', JSON.stringify(user));
+                this.props.history.push('/products')
+            })
+            .catch(response => {
+                console.log(response)
+                this.props.history.push('/register')
+            })
     }
 
     loginClicked = (event) => {
@@ -73,6 +77,7 @@ class Register extends Component {
         let nameValid = this.state.nameValid
         let emailValid = this.state.emailValid;
         let passwordValid = this.state.passwordValid;
+        let confirmPasswordValid = this.state.confirmPasswordValid;
         let addressValid = this.state.addressValid;
         let mobileValid = this.state.mobileValid;
 
@@ -97,6 +102,11 @@ class Register extends Component {
                 passwordValid = value.length >= 8;
                 fieldValidationErrors.password = passwordValid ? '' : ' is too short';
                 break;
+            // Confirmpassword must match password
+            case 'confirmpassword':
+                confirmPasswordValid = this.state.password === this.state.confirmpassword;
+                fieldValidationErrors.confirmpassword = confirmPasswordValid ? '' : ' is invalid';
+                break;
             // Address must be at least length 15
             case 'address':
                 addressValid = value.length >= 15;
@@ -116,15 +126,19 @@ class Register extends Component {
             nameValid: nameValid,
             emailValid: emailValid,
             passwordValid: passwordValid,
+            confirmPasswordValid: confirmPasswordValid,
             addressValid: addressValid,
             mobileValid: mobileValid
         }, this.validateForm);
     }
 
     validateForm() {
-        this.setState({ formValid: this.state.usernameValid && this.state.nameValid 
-            && this.state.emailValid && this.state.passwordValid 
-            && this.state.addressValid && this.state.mobileValid });
+        this.setState({
+            formValid: this.state.usernameValid && this.state.nameValid
+                && this.state.emailValid && this.state.passwordValid
+                && this.state.confirmPasswordValid && this.state.addressValid
+                && this.state.mobileValid
+        });
     }
 
     errorClass(error) {
@@ -153,13 +167,37 @@ class Register extends Component {
                                 <label>Email</label>
                                 <input placeholder="email" name="email" required className="form-control" value={this.state.email} onChange={this.handleUserInput} />
                             </div>
-                            <div className={`form-group ${this.errorClass(this.state.formErrors.password)}`}>
-                                <label>Password</label>
+                            <label>Password</label>
+                            <div className={`input-group ${this.errorClass(this.state.formErrors.password)}`}>
                                 <input placeholder="password" name="password" required className="form-control" type={this.state.showpassword} value={this.state.password} onChange={this.handleUserInput} />
+                                <div class="input-group-append">
+                                    <button class="btn bg-transparent btn-outline-secondary" type="button" onClick={this.showPasswordClicked}>
+                                        {this.state.showpassword === "password" ? <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-eye-slash" viewBox="0 0 16 16">
+                                            <path d="M13.359 11.238C15.06 9.72 16 8 16 8s-3-5.5-8-5.5a7.028 7.028 0 0 0-2.79.588l.77.771A5.944 5.944 0 0 1 8 3.5c2.12 0 3.879 1.168 5.168 2.457A13.134 13.134 0 0 1 14.828 8c-.058.087-.122.183-.195.288-.335.48-.83 1.12-1.465 1.755-.165.165-.337.328-.517.486l.708.709z" />
+                                            <path d="M11.297 9.176a3.5 3.5 0 0 0-4.474-4.474l.823.823a2.5 2.5 0 0 1 2.829 2.829l.822.822zm-2.943 1.299.822.822a3.5 3.5 0 0 1-4.474-4.474l.823.823a2.5 2.5 0 0 0 2.829 2.829z" />
+                                            <path d="M3.35 5.47c-.18.16-.353.322-.518.487A13.134 13.134 0 0 0 1.172 8l.195.288c.335.48.83 1.12 1.465 1.755C4.121 11.332 5.881 12.5 8 12.5c.716 0 1.39-.133 2.02-.36l.77.772A7.029 7.029 0 0 1 8 13.5C3 13.5 0 8 0 8s.939-1.721 2.641-3.238l.708.709zm10.296 8.884-12-12 .708-.708 12 12-.708.708z" />
+                                        </svg> : <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-eye" viewBox="0 0 16 16">
+                                            <path d="M16 8s-3-5.5-8-5.5S0 8 0 8s3 5.5 8 5.5S16 8 16 8zM1.173 8a13.133 13.133 0 0 1 1.66-2.043C4.12 4.668 5.88 3.5 8 3.5c2.12 0 3.879 1.168 5.168 2.457A13.133 13.133 0 0 1 14.828 8c-.058.087-.122.183-.195.288-.335.48-.83 1.12-1.465 1.755C11.879 11.332 10.119 12.5 8 12.5c-2.12 0-3.879-1.168-5.168-2.457A13.134 13.134 0 0 1 1.172 8z" />
+                                            <path d="M8 5.5a2.5 2.5 0 1 0 0 5 2.5 2.5 0 0 0 0-5zM4.5 8a3.5 3.5 0 1 1 7 0 3.5 3.5 0 0 1-7 0z" />
+                                        </svg>}
+                                    </button>
+                                </div>
                             </div>
-                            <div>
-                                <label>Show Password</label>
-                                <input type="checkbox" name="showpassword" className="form-control" onChange={this.showPasswordClicked}></input>
+                            <label>Confirm Password</label>
+                            <div className={`input-group ${this.errorClass(this.state.formErrors.confirmpassword)}`}>
+                                <input placeholder="confirm password" name="confirmpassword" required className="form-control" type={this.state.showpassword} value={this.state.confirmpassword} onChange={this.handleUserInput} />
+                                <div class="input-group-append">
+                                    <button class="btn bg-transparent btn-outline-secondary" type="button" onClick={this.showPasswordClicked}>
+                                        {this.state.showpassword === "password" ? <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-eye-slash" viewBox="0 0 16 16">
+                                            <path d="M13.359 11.238C15.06 9.72 16 8 16 8s-3-5.5-8-5.5a7.028 7.028 0 0 0-2.79.588l.77.771A5.944 5.944 0 0 1 8 3.5c2.12 0 3.879 1.168 5.168 2.457A13.134 13.134 0 0 1 14.828 8c-.058.087-.122.183-.195.288-.335.48-.83 1.12-1.465 1.755-.165.165-.337.328-.517.486l.708.709z" />
+                                            <path d="M11.297 9.176a3.5 3.5 0 0 0-4.474-4.474l.823.823a2.5 2.5 0 0 1 2.829 2.829l.822.822zm-2.943 1.299.822.822a3.5 3.5 0 0 1-4.474-4.474l.823.823a2.5 2.5 0 0 0 2.829 2.829z" />
+                                            <path d="M3.35 5.47c-.18.16-.353.322-.518.487A13.134 13.134 0 0 0 1.172 8l.195.288c.335.48.83 1.12 1.465 1.755C4.121 11.332 5.881 12.5 8 12.5c.716 0 1.39-.133 2.02-.36l.77.772A7.029 7.029 0 0 1 8 13.5C3 13.5 0 8 0 8s.939-1.721 2.641-3.238l.708.709zm10.296 8.884-12-12 .708-.708 12 12-.708.708z" />
+                                        </svg> : <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-eye" viewBox="0 0 16 16">
+                                            <path d="M16 8s-3-5.5-8-5.5S0 8 0 8s3 5.5 8 5.5S16 8 16 8zM1.173 8a13.133 13.133 0 0 1 1.66-2.043C4.12 4.668 5.88 3.5 8 3.5c2.12 0 3.879 1.168 5.168 2.457A13.133 13.133 0 0 1 14.828 8c-.058.087-.122.183-.195.288-.335.48-.83 1.12-1.465 1.755C11.879 11.332 10.119 12.5 8 12.5c-2.12 0-3.879-1.168-5.168-2.457A13.134 13.134 0 0 1 1.172 8z" />
+                                            <path d="M8 5.5a2.5 2.5 0 1 0 0 5 2.5 2.5 0 0 0 0-5zM4.5 8a3.5 3.5 0 1 1 7 0 3.5 3.5 0 0 1-7 0z" />
+                                        </svg>}
+                                    </button>
+                                </div>
                             </div>
                             <div className={`form-group ${this.errorClass(this.state.formErrors.address)}`}>
                                 <label>Home Address</label>
